+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+### Features
+* Use annotations with the prefix `vault.hashicorp.com/alias-metadata-` from the client token's associated
+service account as alias metadata for the Vault entity [GH-226](https://github.com/hashicorp/vault-plugin-auth-kubernetes/pull/226)
+
 ### Changes
 
 * Build with go 1.21.3